requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlmodel==0.0.22
SQLAlchemy==2.0.31
pydantic==2.8.2
pydantic-settings==2.4.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
httpx==0.27.0
python-multipart==0.0.9
pytest==8.2.2
pytest-asyncio==0.23.8
